Describe a platform and its key attributes
This commit brings in the basic concept of a platform. In the case of
vanagon, the platform defines where a package is built, how to turn a
basic instance of the platform into a builder, and how to install build
dependencies. It also includes logic for building the various packaging
artifacts (redhat spec file, debian files, etc).