(RE-3492) Update platform dsl to use methods for make, patch
This commit updates the make and patch attributes to be accessed by a
DSL method rather than the direct instance variable. It also updates the
rpm and deb platform classes to use sane defaults for them so they
should not need to be specified unless they need to be overridden.