(RE-3492) Component DSL updates
This commit updates the component DSL to use build_requires and
requires instead of build_depends_on and depends_on. It also updates the
driver and project to automatically determine if a build_requires entry
is satisfied by a component in the project or not. Any build_requires
not satisfied by components in the project will be fetched from the
external source determined by the platform. requires are reserved for
external runtime dependencies at package install time.