(RE-3489) Add service handling to the component
This commit adds an install_service method to the component DSL. This
method takes a default file, service file and name and registers it to
the component. It appends service installation to the install steps and
adds the service files to the extra files list. Accessors are added for
service and files.