(RE-3492) Remove to_codename method
This commit removes the to_codename method from the platform class,
because it is quite brittle and requires a fair amount of maintenance.
Instead, platforms will be responsible to declaring their own codename
if they have one. This also updates the changelog template to directly
reference the codename attribute instead of calling to_codename.
Platforms that need codenames defined will need to be updated
accordingly.