(RE-3493) Extract source into separate class
For the support of git sources, a first step is to extract the http
source support into its own class. This commit accomplishes that by
creating a base class and moving several relevant utilities into the new
http class. The main component class is also updated to reference the
new source class.