(RE-3493) Add git support to component sources
Now that source has been abstracted into http and git, this commit adds
the git class for git support. It implements most of the same methods as
the http source class, even though some of them are noops. The ref
component dsl method is added to delineate the sha/tag/branch to check
out from the referenced git repo.