Add getter to the platform dsl
Because the object in the platform block is actually an instance of the
platform dsl and not the platform itself, attempt to use attributes like
platform.servicetype will fail, as it is a method on the dsl. This
commit adds a method_missing catch to the dsl to redirect get_* methods
to return the attribute of the platform class. So
platform.get_servicetype will return the servicetype for the platform.