(RE-3493) Add json parsing for the component dsl
This commit adds a simple dsl method to parse the contents of a json
file and use it to run some component dsl. This will facilitate CI for
the AIO, allowing a json blob for a given puppet/facter/hiera/mco build
to be dropped into the project, and that json will specify the git
source and ref to use for the build.