(RE-3596) Update packaging output layout
This commit updates the layout used for storing built packages to match
the conventions set by the packaging repo. This should enable the
leveraging of its internals to ship vanagon projects. The output_dir
method is added to both deb and rpm platforms and it is used in the
generate_package method to ensure that the packages end up in the
expected location. The constructors for rpm and deb are also updated to
invoke the parent constructor to ensure that os_name and os_version are
set.