(RE-3591) Update packaging to use project directories
Previously, the rpm and deb packaging was using hardcoded directories of
etc, opt and var. This isn't very flexible and makes some assumptions
that won't always hold true. This commit updates the packaging to use
the directories declared in the project config to guide what is
installed. The get_directories method is added to the project class to
facilitate these changes.