(RE-3661) Provide number of available processes
In some cases, we would like to build projects in parallel, when
available. When using make, we can specify the number of jobs to run in
parallel. This number usually corresponds to the number of processes
available on the machine plus one. This pull request adds in the ability
to determine how many processes are available for use, and makes that
data available during build time.