(RE-3685) Add repo binary to bin to make repos
Now that artifacts are shipping for the composite artifact, we need to
be able to make repos of them for consumption by beaker. This commit
adds a repo binary which makes repos of the shipped packages. This
assumes the packages have been built and shipped, and assumes both deb
and rpm packages have been built.