(RE-3688) Add configfile component dsl method
Previously, there was no way of designating configfiles in vanagon,
which means that config files would just be overwritten on upgrade,
which is bad bad bad. This commit adds 2 mew component dsl methods. The
first is configfile, which allows the user to mark an already installed
file as a configfile. The second combines install_file and configfile to
allow a user to install a file and designate it a configfile in one
method call.