(RE-3824) Allow a vanagon project to have runtime deps in package
Prior to this commit, there was not a way to specify a run-time
dependency in a package generated by vanagon. Each component could have
a dependency for build time, but that didn't help much.

This was created because code-management needed to depend upon the
puppet-agent both at build and at run time.

This is implemented by adding a requires parameter to the main project
object.

Limitations: requires is a single string, and is defined at a project
level and is not configurable per OS type (e.g debian has one name and
el has another). Thus far that is an ok trade-off because the only
dependency has the same name. That may need to change in the future.