(RE-3824) Allow a vanagon project to have runtime deps in package
Prior to this commit, there was not a way to specify a run-time
dependency in a package generated by vanagon. Each component could have
a dependency for build time, but that didn't help much. There was a
component property called requires, but it wasn't used in any packaging.

This was created because code-management needed to depend upon the
puppet-agent both at build and at run time.

This is implemented by adding a requires parameter to the main project
object as well as bubbling up each component-level requires into the RPM
spec file and debian control file.

Requirements are then flattened via the get_requires method found in
project.rb.

Note: If you have different requirements on different platforms, you're
probably going to want to mess with that in the component file and not
the project-level one. The project level one is primarily for uniform
dependencies or our own generated metapackages.