(RE-3828) Present an error message when a non 200 happens
Previously, when vanagon couldn't download a source file vanagon would
just throw a checksumming error rather than tell you that it couldn't
find the payload requested. This simple change let's you know if the
response is not a 200.

Note: this doesn't honor redirects, but we also don't serve them in the
current setup.