(RE-3856) Component listing in pkg desc should be one per line
Previously each component in a vangon-built package was listed in
arbitrary order and on one line comma separated. This patch changes taht
behavior to make it one component per line and orders them
alphabetically. This should make it easier to parse the output of
querying the package with grep or other such tools.