(RE-3815) Lay down FS level diffs during build
In order to only package the files that were created during the build,
this commit lays the groundwork by dropping three files via the
makefile. The first file is the files in the directories owned by the
project before the build, the second is the files after the build, and
the third is the difference between the two (files added by the build to
those directories).