(RE-3815) Use the file-list to drive rpm and deb builds
This commit leverages the file-list laid down during the build to decide
what to put in the tarball and include in the package. The file-list is
concatenated onto the install file for debian and used via the -f flag
to the %files list in the rpm specfile. Directories are no longer copied
wholesale as they may have files that shouldn't be copied.