(RE-3831) Add helpers to provision package repositories
This is a patch to add feature and syntax to platform setups that
provision yum or apt repositories. This is done by using

platform.yum_repo
or
platform.apt_repo

in the configs/platform definition files.

The dsl usage looks like

platform.yum_repo "repo-definition", "repo title"

repo-definition can be a uri that is text and contains the text for a
yum/apt repo, or can be a package, such as a puppetlabs-release package
in native format (deb/rpm).

This should replace the shell overloading in platform.provision.

You can call platform.yum_repo (or apt_repo) N times.