(RE-3719) Add Directory class to vanagon
As both components and projects may be adding directories with the same
properties, it makes sense to abstract it into a light class. This
commit adds a Vanagon::Common::Directory class to allow projects and
components to add directories with specific permissions and ownership.
Current packaging is updated to maintain the status quo.