(RE-3719) Apply mode, owner, group to directories in packaging
This commit updates the packaging to apply any overrides to the default
mode, owner or group to the project directories in the deb and rpm
packaging. In the deb packaging, this is accomplished via the postinst,
in rpm via the %attr macro in the specfile. The directory dsl method for
projects is updated to take mode, owner and group via keyword arguments,
which moves the required ruby version to run vanagon to 2.0.