(RE-3719) Update debian postinst to warn if user/group doesn't exist
If the specified user/group doesn't exist at package install time, there
isn't much we can do about it except warn the user and keep on going.
This commit adds a warning in the case that the desired user/group for a
directory does not exist, so that the package installation can still
succeed.