(RE-3951) Update templates, project and component to use pathname
This commit updates the install_file and configfile component dsl
methods to add Pathname objects to the collection instead of just the
location of the file on disk. Tests are updated to reflect this and the
templates are updated to extract the path from the pathname to ensure
that packaging continues to behave as before.