(maint) Update specs to use include, remove unneeded parameter
This commit updates the specs to use the include matcher instead of
be_include, which are functionally equivalent. include is built-in to
rspec. It also updates the component dsl tests to pass in a platform
object into the component constructor only when needed. This was likely
copypastad during the initial creation of the tests.