(RE-3967) Collect flags and args in bin/build
Previously a combination of array position and array inclusion were used
to get both arguments and flags. This broke grabbing things like
--preserve out of the command line because it would interpret --preserve
as the target, or platform, or project, which wouldn't work. This commit
updates the script to collect flags and args separately from ARGV. This
will allow the flags to be collected without affecting other arguments.