(RE-3970) Add logging of host usage with VMpooler
When using vanagon rapidly and prototyping, you'll end up grabbing lots
of hosts from the VMpooler. By default, hosts are destroyed if vanagon
completes sucessfully, but when it fails, the hosts is kept around. When
working on lots of projects, that can leave many orphans that only get
cleaned up by the VMpooler reaping.

This adds a log in the local directory that shows what hosts are grabbed
(and what pool they are from). It will also show when they are
destroyed.

    # Logfile created on 2015-02-08 05:13:11 +0000 by logger.rb/41954
    I, [2015-02-08T05:13:11.315778 #29183]  INFO -- vanagon: Reserving nand9d4v3tkcrzu (sles-10-x86_64)
    I, [2015-02-08T05:22:48.501426 #29183]  INFO -- vanagon: nand9d4v3tkcrzu has been destroyed