(RE-3827) Rescue common Net::Http exceptions
For the http source, previously net/http exceptions were not rescued at
all, which meant that figuring out what had gone wrong was difficult.
This commit adds some exception handling to the download method in the
http source. It allows a bit of context to be added before reraising the
exception.