(RE-4064) Add devkit command to build specific component
Adds a new command, `devkit`, that creates a dev environment for
a project - or if components are listed, specific components in a
project - and builds and installs it. Unlike the `build` command, it
always preserves remote machines and stops before packaging.

Makes workdir and configdir configurable for both `devkit` and `build`
commands.