(maint) Add retry_with_timeout method to utilities
In the docker engine, we currently have a sleep rather than doing the
right thing, which is polling on the host and port for a given time. The
sleep will invariably fail on a slow enough host. This commit adds a
utility method to allow a set number of retries of a given block, which
can be used by docker to poll using ssh against the container.