(RE-3822) Add support for build engines to vanagon
Previously target selection, setup and teardown was written into
driver.rb, with some conditionals to decide which path to take. This
commit tries to make this situation better by adding the concept of an
engine to vanagon. The engine will define how to select a build target,
how to set it up as a builder, and how to tear it down when done. The
initial work includes a base engine which supports ssh, and a docker and
pooler backend.