(RE-3822) Refactor driver.rb with engine support
Now that there are engines, driver.rb can be refactored to remove the
logic that switched between different behaviors. This can be entirely
driven by the engine. This commit begins passing an engine_name into the
driver class, which then loads it and uses it to select a build target,
set it up, and tear it down after a build.