(RE-3822) Refactor driver.rb and build
This commit simplifies the glue between build and driver. Previously,
the Vanagon class simply acted as a direct pass through to a
Vanagon::Driver method of the same name. This was a bad idea. This
commit replumbs that interface so that build just directly instantiates
the Vanagon::Driver class.
Additionally, the load_* methods are removed from driver.rb and moved
inline into the initializer, instead of being invoked from run.