(maint) Update repo binary to take an optional argument
Because the packaging repo tends to fail when only one platform is
available, this commit updates the repo binary to take an optional
argument of repo target. If deb or rpm are passed as the argument, only
that set of repos will be created. If no arguments are passed (or
garbage), both will be created. This allows a jenkins jobs to decide
which repos to create based on the build targets.