(RE-4235) Add target_repo method to project DSL
In order to ship to alternate repositories, the initial packaging layout
must first match the layout expected by packaging. This commit adds a
target_repo method to the vanagon DSL to allow this. When set the built
artifacts will be laid out with the alternate repo in the same location
as currently done in packaging.