(RE-4055) Add provides support to component DSL
It is useful for a package to be able to specify that it provides a
capability. This was previously not possible in vanagon, but this commit
adds a provides dsl method to the component which allows components to
express provides that can be collected for the project and package. In
rpm and deb packages, this is expressed as a provides in the spec or
control file.