(RE-4159) Add versioned provides and replaces to vanagon
Using unversioned provides and replaces works on modern versions of rpm,
but falls down on EL5 and earlier. It is also not a best practice in
packaging. This commit adds the ability to version provides and
replaces, and updates the packaging accordingly. Debian packaging
doesn't support versioned provides, so that remains unversioned.