(maint) Allow rpmbuild command to be specified per platform
You'd think that rpmbuild works the same on all platforms, well you'd
pretty much be right unless you time-travelled to 1999 and fork RPM and
put it on AIX. Then you have to go back to the days when rpm was only
one command you just passed '-bb' to rpm to build stuff. It's like
RHEL3, but worse.

This commit just plumbs the ability to specify rpmbuild command with
arguments and sets a sane default for all platforms. You'll need to
specify the AIX rpmbuild command inside its platform definition.