(RE-4254) Fix yum_repo helper for EL5
When using EL5, the yum_repo helper doesn't work against an rpm target
because yum is just too darn old to do a yum localinstall against an
http target. You must first download the rpm, then install it.

This could have been implemented with just a raw call to rpm, but since
curl has a better ability to work with http_proxy environment variables,
and a curlrc, the safer thing seems to be to download the package,
install it and then remove it. This is also how the helpers for zypper
and apt are implemented, so for congruity, we'll keep it that way.