(RE-4097) Explicitly call bash
Shell expansions works differently in different shells. Prior to this
commit, there was an implied assumption that when executing remote
commands over ssh, bash (or something bash compatible) would be the
shell. That doesn't hold true on traditional unix where sh or ksh is the
default shell when executing remote commands over ssh.

This commit changes the call to be 'bash -c' and then the command,
rather than just assuming bash.