(RE-4257) Add more verbose output for component source fetching
Previously when fetching a git source you would need to guess as to what
the ref was and would have no idea what url was being cloned. This
commit adds some basic notifications to the build of what url is being
fetched, and what ref for git, or what file for http sources.