(RE-4374) Add noarch project dsl method
Previously vanagon assumed all projects would be arch specific. This is
not the case, so this commit adds a noarch dsl method to vanagon
projects where they can declare that they are noarch. Noarch package
templates have the correct bits set to trigger noarch builds and the
expected package names are updated as well.