(RE-4381) Update rpm spec to not run python byte compile
Prior to this commit, we were adhering to the normal RPM convention of
letting the package byte-compile python files for optimizations at
runtime. This isn't a terrible idea, however it is incompatible with the
way vanagon calculates file-lists and deltas to see what should be
included in the package.

This patch simply disables the call to the python-byte-compile brp
script.