(RE-3978) Set known hosts file to /dev/null
When using the pooler, over a long enough time period, a host will be
provisioned with the same ip, but a different hostname (or vice versa).
This causes ssh and rsync to present an interactive prompt, which can
break automated builds. This commit adds an option to all ssh and rsync
calls to set the known hosts file to /dev/null to avoid this problem.
/dev/null will never have any hosts listed or stored.