(RE-4256) Add component bill-of-materials to packages
This commit adds a bill-of-materials of all of the components and
version to the package being produced. It is included as a doc in both
the rpm and deb packages. The previous method listing the versions in
the package metadata has been refactored as well so that the behavior is
consistent.