(maint) Update devkit for recent makefile changes
A recent change to vanagon updated the makefile to allow components and
projects to share the same name without having 2 targets with the same
name in the makefile. This unfortunately broke the devkit workflow,
which assumes a target with the name of the project will exist in the
makefile. This commit updates the call to make to correspond to the
changes made, appending '-project' to the target.