(RE-4035) Add url rewrite support to vanagon
This commit adds some rewrite support to vanagon sources. A rewrite rule
can be added at the project level for one of the supported source types
(http, git, file) and it will be applied to all source urls of that
type. In the simple case, string substitution is applied to the url to
replace the target host with a new one, and in the complex case, a Proc
is passed the url to apply custom logic to.