(maint) Allow application of multiple patches
The way the patch application was written, it correctly apply the first
patch before leaving the cd block and ascending to the parent directory,
at which point any subsequent patch applications would fail miserably.
This commit corrects that by ensuring that the patch applications all
happen within the context of the cd.