(RE-4588) Allow users to pass in the gpg key for provisioning an apt repo
Prior to this commit, when provisioning an apt repo, a user could pass
in either a .list file or a .deb file, which would be curled down, and
installed into the right location. The only issue with this was when the
.list file was installed, if the gpg key for that repo wasn't already on
the system, no packages from that repo could be installed.

This commit updates the logic of the apt_repo method to allow the user
to pass in an optional gpg key argument to download and install during
the provisioning process. This also pulls out the logic to allow a user
to name the .list file they are installing as this is meant to happen on
a temporary machine, and users do not need that much control.