(RE-4095) Add get_root_directories method to project
For solaris 11, it will be useful to have a list of the directories
owned by the project that are closest to /. For example, if a project
owns /a/b/c and /a/b, we would be interested only in /a/b. This commit
adds such a method to the project class.