(RE-4594) Fall back to older `find` syntax
Currently, `find -L` fails when run on el4. On this platform, we only
have find 4.1 available, which doesn't support the newer `-L` flag. When
we try to pass in this flag to find, it fails, and subsequently vanagon
builds an empty package for el4. This commit updates vanagon to fall
back to the old find syntax to allow us to deal with symlinks on el4 and
other systems that only have old versions of find available