(RE-4095) Include service_file in the services defined in components
With the addition of the service_type for smf services, services can no
longer be easily defined by combining the service dir and service name.
This commit addresses that deficiency by including the full path to the
service file in the service attribute attached to the component. This
requires updating all uses of @component.service that assume it will be
a string to reference the name attribute of the service.