(RE-4095) Handle solaris 10 'configfiles' in a sane manner
Solaris 10 packaging has absolutely no concept of a configfile. There
are just files to be installed. Our previous packaging handled this by
putting the current configfiles in a pristine directory and copying them
into place if the configfile was absent. This commit updates the
configfile component DSL method to take a similar approach. For solaris
10, we move any configfile in the package to add a .pristine suffix, and
then use a postinst script to move those files into place if the target
doesn't exist.