(RE-4095) Add optional suffix to build dependency installation
On some platforms, the build dependency installation may return a
non-zero exit code (I'm looking at you solaris 11). In those cases
adding a command or shell snippet to append to the installation command
is essential. This commit adds an optional parameter to the
build_dependencies platform dsl method to enable this behavior.