(RE-4095) Add Solaris 10 support to vanagon
This commit adds Solaris 10 as a target and supported platform type to
vanagon and also adds an is_solaris? helper method to the platform
class. This work is largely based on the automation used in
enterprise-dist. Solaris 10 and 11 are different enough that they make
sense to model in their own classes.