(RE-4550) Use a buildroot other than $HOME
Previously vanagon would simply use root's homedir to build in, which
works most of the time. On some systems, root's homedir is /, which can
make debugging build failures harder. This commit updates the base
engine to get and set a remote workdir to store the build artifacts and
to run the build from.