(RE-4158) Vanagon user management is now slightly less bad
Vanagon was previously passing --system to usermod/groupmod calls,
and that's not going to work. Now, through some logistic gyrations
and Array-to-String abuse, usermod/groupmod are safe again!

(To be clear, I don't like this solution but I don't want to break
the interface for any projects using this in an attempt to refactor
it into something more palatable)