(RE-4610) Enable additional cleanup during builds
Some projects are large and can fill up the buildroot on systems with a
small amount of disk allocated to tmp. This commit adds a directive
which can be used in project configs to enable a more agressive cleanup
during builds (more agressive than the current nothing). Source archives
are deleted after unpacking, and the unpacked source is deleted after
the project has been fully installed on the system.