(RE-4860) Add write_version_file method to Project DSL
This commit adds a simple method to the project DSL to facilitate
writing out the version of the project to a file in the package
generated by vanagon. The makefile is also updated with a conditional
target to do the work if the attribute is set. Oh and there are tests.