(RE-4897) Use correct dist macro definition in rpm_defines
Previously we incorrectly specified the dist macro. All --defines passed
along the command line are accessed using % in the spec file, but
defined without it on the command line. This commit removes the % so
that the dist macro can be correctly defined for platforms that don't
already have it available (el4, el5, sles10).