(RE-4606) Add support for environment overrides in components
Currently, in vanagon projects, if a specific environment is required
for a component, the convention is to prepend the configure, build and
install steps with the environment variables. This repetition is
undesirable, so this commit implements an environment component dsl
method which takes a hash and exports each key/value pair in the
Makefile template for configure, build and install.