(RE-5032,5089) OSX pre post install script tweaks
This commit addresses :

* osx puppet-agent clobbers config files if they already exist

Using the same logic as Solaris 10, we append a "pristine" suffix to
config files and only move them into place if they do not exist.

* puppet and mco should not be started after osx puppet-agent install

Currently, the puppet-agent installer starts both puppet and mco post
install.  This isn't the desired behavior.  The new behavior only starts
the processes if it detects we're upgrading ( by checking for existence
of a package receipt ).