(RE-4607) Allow a more flexible build dependency installation
Previously platforms were limited by a single prefix and suffix to
install build dependencies, which is insufficient for some platforms
that don't have sane package managers or have multiple methods of
package installation. This commit addresses that by allowing a fallback
method that is baked into the platform class itself. The platform can
define a method that takes a list of packages to install and does the
right thing with them.