(maint) Update num_cores in solaris 10 for Makefile safety
The previous fix to the num_cores attribute for solaris 10 neglected to
take into account that it would only be used in makefiles. This means
that without 2 $ signs, make would attempt to interpolate a variable
called $1 in the awk call (and something similar in the grep. This would
cause make to spew errors about invalid expr, as the call would fail. This
commit addresses this problem by using two $ signs in the command.