(maint) Add empty directories to solaris 10 packages
This unfortunate commit is brought to us by svr4. Previously empty
directories were not being added to the solaris 10 package because they
wouldn't be in the tarball. Adding them to the proto.erb is challenging
because duplicate entries are disallowed, and the ruby layer won't know
in advance what directories will be in the package by virtue of having
files under it. This commit works around these problems by adding any
directory entries that aren't present in the proto to the proto (using
ggrep to check).

I am so sorry.