(RE-5251) Refresh apt cache upon vanagon build start
Prior to this, if the apt cache was stale for a template, vanagon would
fail to install curl, and then move onto to attempting to do other
things. This commit ensures that the apt cache is updated as the first
operation, which should prevent this type of error in the future.