(maint) Prevent duplicate entries from file-lists
When directories that contain each other are declared in vanagon the
find calls to generate file lists will have 2 copies of any files in the
deepest directory. This causes problems with the tar on el4 which will
prevent a successful package build. This commit works around that
problem by piping the find | sort to uniq. Uniq should be available on
all systems that vanagon currently supports and any unix/linux variety.