(RE-5290) Sources should be optional in a component
Sometimes it is useful to have a component that does some work, but has
no explicit source. This was previously impossible due to the way
get_source worked, by always instantiating a source. get_source now only
instantiates a source if url is present, and emits a warning for
components without a source.