(RE-5374) Define eql? and hash for Pathname objects
Previously, Pathname objects did not define hash or eql?, which means
that despite a and b being equal according to the == method, calls to
uniq or additions to a set would add the same object multiple times.
This commit defines both eql? and hash to avoid this problem. With this
commit adding the same pathname object to a set will result in no
addition and uniqing an array with the same object multiple times will
remove duplicates.