(RE-5374) Only include files, directories once in list
Previously if a directory or file was included multiple times in a
component or project, it would be listed multiple times. This causes
problems with certain package managers and is also unnecessary. This
commit avoids this problem by calling uniq on the arrays before
returning them, so no duplicates are included.