(RE-5374) Limit extra directory collection to wr_linux
Previously the adding of directories to the rpm files list was applied
to all builds. This has the unfortunate consequence of breaking aix
builds, in which duplicate file declarations is a hard error. This
commit limits that collection to just wr_linux, which has rpm5, which
cares about this. This also adds "" around the pathnames, to ensure
consistency with the existing file list.