(RE-5380) Add a changelog to vanagon
In order to allow consumers of vanagon to decide if they want to opt
into a new release of vanagon, it should have a change log to show what
has changed between versions. This commit adds a change log formatted
according to http://keepachangelog.com/. It only includes new entries
and does not attempt to recreate past versions' change logs.