(RE-5391) Escape transformations in IPS packaging
Previously, the IPS manifest would apply transformations to paths which
could contain special regex characters without escaping them. Because
the path in the manifest is a regex, these characters must be escaped
for builds with files such as libstdc++.so to succeed. This commit
updates all of the paths to include a regexp escape call.