(RE-5381) Add add_build_repository platform dsl method
This commit adds a generic add_build_repository method to the platform
DSL. This allows a user to not need to know the platform specific name
of the method, and to have a common interface. It also pushes the
platform specific knowledge into the platform class, which is
desireable.
It also includes the first use of the method, the solaris 11 repository.