(RE-5300) Add all file types to file-lists
Previously the files added to the file-lists were only files. This
unfortunately meant that empty directories would not be included in
tarballs or packages. This commit works around this by removing the
restriction from the find call.
This should be safe in all cases. In the case where a directory is added
during the build, the find would already collect all files in that
directory after the build, so there is no change. In the case where a
directory is present before the build (as in a build dependency), it
will still be filtered out after the build.