(maint) add support for platforms that cannot remotely install build deps
For some platforms (HuaweiOS for example), the rpm command is compiled
without support for installing packages via remote uris for security
reasons. So for these platforms we create a tmpdir and download the
build dependency packages to it using curl, and then issue the package
installation command against the locally downloaded files.