Allow for VMPOOL_TOKEN to be used
Other people are already using VMPOOL_TOKEN, we should allow that
instead of making them change the env var they use.

I realize that people seem to think VMPOOLER_TOKEN is more correct, and
that's fine. We don't have to advertise this option, but I'd rather it
not break people just because.