(RE-5627) Install configfiles explicitly on debian
In debian, the conffiles file only declares conf files, it does not
install them. They must be explicitly installed in the debian/install or
debian/rules files. This commit adds configfiles to the install
template. They were previously included incidentally by the behavior of
configfiles in vanagon, which was recently fixed.