(RE-5314) Set -T instead of -t for remote_ssh_command
remote_ssh_command was largely copied from the packaging repo, which
needs to force a tty because jenkins jobs often ssh out to other
systems, which can in turn ssh out to other systems. If a tty is not
present, some commands will fail. Vanagon doesn't leverage any
interactive commands that would require a tty, so explicitly disallowing
ttys in its ssh calls will be safe. This has the added bonus of
unsetting TERM in the remote environment, which has caused problems on
certain platforms.