(maint) Finish build deps before configuring dependents
Previously, a build_requires step would only cause the final, top-level
target of the dependent to have a dependency. This means that it's
possible for the entire dependent to be built before the dependency,
entirely defeating the purpose of build dependencies.

Now, the dependent-configure target depends on the dependency target,
which ensures that include directories, etc will be properly found
when the dependent is configured.