(RE-5565) Only check response for ok in select_target
Having select_target check that response is non-nil/truthy before
checking to see if it 'ok' is in the hash doesn't serve a purpose. Any
failures in setting response will raise an error from the http_request
method, and {} is truthy (and won't cause a problem when looking for
'ok'). This resolves a PerceivedComplexity violation in rubocop.