(RE-5565) Refactor apt_repo for PerceivedComplexity
This commit aims to make the debian build repo logic more manageable.
This happens here in two steps. First, apt_repo is deprecated in favor
of add_build_repository, which is a generic pass through to the platform
method. Second, the method is broken down into smaller chunks, such as
add_gpg_key and add_repository which are then invoked from the main
method.
The provision_with method is also updated here to flatten the resulting
array, as some of the debian methods return an array of strings.