(CPR-240) Ensure AIX doesn't include newlines when creating puppet subsystem
Previously, when reading in the puppet and mcollective service
scripts to create their subsystems on AIX, a newline in the string
was causing the command and its arguments to appear on separate lines.
This caused `mksys` to not set signorm and sigforce, rendering
the service unstoppable.

We now chomp the newline to ensure the arguments are included with
the service command.