(RE-5565) Remove check from remote_ssh_command
remote_ssh_command was largely borrowed from the packaging repo, where
it was used frequently in different contexts and needed stronger
safeguards than in vanagon. This command is only used in a few cases,
and in all uses, target will never be nil or false, which is the only
time this check will be useful. As the parameter has no default, the
caller would need to pass nil/false in intentionally, which seems
unlikely.